requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.5
Flask-SQLAlchemy==1.0
Flask-Testing==0.4
Jinja2==2.7.1
Markdown==2.3.1
MarkupSafe==0.18
Pygments==1.6
SQLAlchemy==0.8.2
Sphinx==1.2b1
Werkzeug==0.9.4
docutils==0.11
gunicorn==18.0
itsdangerous==0.23
nose==1.3.0
psycopg2==2.5.1
pyasn1==0.1.7
pycrypto==2.6
pytz==2013b
requests==1.2.3
rsa==3.1.1
six==1.4.1
wsgiref==0.1.2
